Refactor footer section
