+++ src/components/Footer/LinksWrapper.jsx
@@ -1,0 +1,7 @@
+import styled from "styled-components"
+
+export const LinksWrapper = styled.div`
+  padding: 1.2rem 2.4rem;
+`
+
+export default LinksWrapper
